Hub area
A neutral area where the player can take a break, and in the future save items?

Player starts in this area, and has the option to return after each boss by choosing the door to go through.

Only one door will appear instead of four, or two if its the end of a boss battle. Door position is random